*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@51 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/prim/simple_polygon_3d.inl
@@ -462,7 +462,7 @@ const bool SimplePolygon3D<T>::contains(const TPoint& iP) const
 template <typename T>
 const bool SimplePolygon3D<T>::isInRange(int iIndexOfVertex) const
 {
-	return iIndexOfVertex >= 0 && iIndexOfVertex < vertices_.size();
+	return iIndexOfVertex >= 0 && iIndexOfVertex < static_cast<int>(vertices_.size());
 }
